Remove Cardano.Config.Types and remove unused cli commands
