+++ cardano-cli/src/Cardano/CLI/Byron/Delegation.hs
@@ -23,13 +23,13 @@ import           Cardano.Binary (Annotated (..), serialize')
 import qualified Cardano.Chain.Delegation as Dlg
 import           Cardano.Chain.Slotting (EpochNumber)
 import qualified Cardano.CLI.Legacy.Byron as Legacy
-import           Cardano.Config.Types (CertificateFile (..))
 import           Cardano.Crypto (ProtocolMagicId, SigningKey)
 import qualified Cardano.Crypto as Crypto
 
 import           Cardano.CLI.Byron.Key (ByronKeyFailure, CardanoEra (..), renderByronKeyFailure,
                      serialiseSigningKey)
 import           Cardano.CLI.Helpers (textShow)
+import           Cardano.CLI.Types (CertificateFile(..))
 
 data ByronDelegationError
   = CertificateValidationErrors !FilePath ![Text]
