+++ cardano-cli/src/Cardano/CLI/Helpers.hs
@@ -28,7 +28,7 @@ import           Cardano.Chain.Block (fromCBORABlockOrBoundary)
 import qualified Cardano.Chain.Delegation as Delegation
 import qualified Cardano.Chain.Update as Update
 import qualified Cardano.Chain.UTxO as UTxO
-import           Cardano.Config.Types
+import           Cardano.CLI.Types
 
 
 data HelpersError
