+++ cardano-cli/src/Cardano/CLI/Shelley/Commands.hs
@@ -10,8 +10,6 @@ module Cardano.CLI.Shelley.Commands
   , NodeCmd (..)
   , PoolCmd (..)
   , QueryCmd (..)
-  , BlockCmd (..)
-  , SystemCmd (..)
   , GovernanceCmd (..)
   , GenesisCmd (..)
   , TextViewCmd (..)
@@ -29,24 +27,18 @@ module Cardano.CLI.Shelley.Commands
   , OpCertCounterFile (..)
   , OutputFile (..)
   , ProtocolParamsFile (..)
-  , SigningKeyFile (..)
   , WitnessFile (..)
   , TxBodyFile (..)
   , TxFile (..)
-  , VerificationKeyFile (..)
   , VerificationKeyBase64 (..)
   , GenesisKeyFile (..)
   , MetaDataFile (..)
   , PoolId (..)
   , PoolMetaDataFile (..)
-  , GenesisFile (..)
   , PrivKeyFile (..)
   , BlockId (..)
-  , QueryFilter (..)
-  , StakePoolVerificationKeyHashOrFile (..)
   ) where
 
-import           Data.Set (Set)
 import           Data.Text (Text)
 import           Prelude
 
@@ -55,8 +47,8 @@ import           Cardano.Api.Typed hiding (PoolId)
 
 import           Ouroboros.Consensus.BlockchainTime (SystemStart (..))
 
-import           Cardano.Config.Types (CertificateFile (..), NodeAddress, SigningKeyFile (..),
-                     UpdateProposalFile (..))
+import           Cardano.CLI.Types
+
 import           Shelley.Spec.Ledger.TxData (MIRPot)
 
 --
@@ -73,8 +65,6 @@ data ShelleyCommand
   | NodeCmd         NodeCmd
   | PoolCmd         PoolCmd
   | QueryCmd        QueryCmd
-  | BlockCmd        BlockCmd
-  | SystemCmd       SystemCmd
   | GovernanceCmd   GovernanceCmd
   | GenesisCmd      GenesisCmd
   | TextViewCmd     TextViewCmd
@@ -93,9 +83,6 @@ data StakeAddressCmd
   = StakeAddressKeyGen VerificationKeyFile SigningKeyFile
   | StakeAddressKeyHash VerificationKeyFile (Maybe OutputFile)
   | StakeAddressBuild VerificationKeyFile NetworkId (Maybe OutputFile)
-  | StakeKeyRegister PrivKeyFile NodeAddress
-  | StakeKeyDelegate PrivKeyFile PoolId Lovelace NodeAddress
-  | StakeKeyDeRegister PrivKeyFile NodeAddress
   | StakeKeyRegistrationCert VerificationKeyFile OutputFile
   | StakeKeyDelegationCert VerificationKeyFile StakePoolVerificationKeyHashOrFile OutputFile
   | StakeKeyDeRegistrationCert VerificationKeyFile OutputFile
@@ -125,7 +112,6 @@ data TransactionCmd
   | TxSign TxBodyFile [SigningKeyFile] (Maybe NetworkId) TxFile
   | TxWitness TxBodyFile SigningKeyFile (Maybe NetworkId) OutputFile
   | TxSignWitness TxBodyFile [WitnessFile] OutputFile
-  | TxCheck         -- { transaction :: Transaction, nodeAddr :: NodeAddress }
   | TxSubmit Protocol NetworkId FilePath
   | TxCalculateMinFee
       TxBodyFile
@@ -150,10 +136,7 @@ data NodeCmd
   deriving (Eq, Show)
 
 data PoolCmd
-  = PoolRegister PoolId   -- { operator :: PubKey, owner :: [PubKey], kes :: PubKey, vrf :: PubKey, rewards :: PubKey, cost :: Lovelace, margin :: Margin, nodeAddr :: NodeAddress }
-  | PoolReRegister PoolId -- { operator :: PubKey, owner :: [PubKey], kes :: PubKey, vrf :: PubKey, rewards :: PubKey, cost :: Lovelace, margin :: Margin, nodeAddr :: NodeAddress }
-  | PoolRetire PoolId EpochNo NodeAddress
-  | PoolRegistrationCert
+  = PoolRegistrationCert
       VerificationKeyFile
       -- ^ Stake pool verification key.
       VerificationKeyFile
@@ -185,43 +168,26 @@ data PoolCmd
   deriving (Eq, Show)
 
 
-data QueryCmd
-  = QueryPoolId NodeAddress
-  | QueryProtocolParameters Protocol NetworkId (Maybe OutputFile)
+data QueryCmd =
+    QueryProtocolParameters Protocol NetworkId (Maybe OutputFile)
   | QueryTip Protocol NetworkId (Maybe OutputFile)
   | QueryStakeDistribution Protocol NetworkId (Maybe OutputFile)
   | QueryStakeAddressInfo Protocol StakeAddress NetworkId (Maybe OutputFile)
   | QueryUTxO Protocol QueryFilter NetworkId (Maybe OutputFile)
-  | QueryVersion NodeAddress
   | QueryLedgerState Protocol NetworkId (Maybe OutputFile)
-  | QueryStatus NodeAddress
   deriving (Eq, Show)
-
-
-data BlockCmd
-  = BlockInfo BlockId NodeAddress
-  deriving (Eq, Show)
-
 
 data GovernanceCmd
   = GovernanceMIRCertificate MIRPot [VerificationKeyFile] [Lovelace] OutputFile
-  | GovernanceProtocolUpdate SigningKeyFile -- { parameters :: ProtocolParams, nodeAddr :: NodeAddress }
   | GovernanceUpdateProposal OutputFile EpochNo
                              [VerificationKeyFile]
                              ProtocolParametersUpdate
-  | GovernanceColdKeys SigningKeyFile     -- { genesis :: GenesisKeyFile, keys :: [PubKey], nodeAddr :: NodeAddress }
   deriving (Eq, Show)
 
 
 data TextViewCmd
   = TextViewInfo !FilePath (Maybe OutputFile)
   deriving (Eq, Show)
-
-data SystemCmd
-  = SysStart GenesisFile NodeAddress
-  | SysStop NodeAddress
-  deriving (Eq, Show)
-
 
 data GenesisCmd
   = GenesisCreate GenesisDir Word Word (Maybe SystemStart) (Maybe Lovelace) NetworkId
@@ -261,10 +227,6 @@ newtype TxByronWitnessCount
 
 newtype BlockId
   = BlockId String -- Probably not a String
-  deriving (Eq, Show)
-
-newtype GenesisFile
-  = GenesisFile FilePath
   deriving (Eq, Show)
 
 newtype GenesisKeyFile
@@ -336,23 +298,7 @@ newtype TxFile
   = TxFile FilePath
   deriving (Eq, Show)
 
-newtype VerificationKeyFile
-  = VerificationKeyFile FilePath
-  deriving (Eq, Show)
-
 -- | A raw verification key given in Base64, and decoded into a ByteString.
 newtype VerificationKeyBase64
   = VerificationKeyBase64 String
-  deriving (Eq, Show)
-
--- | UTxO query filtering options.
-data QueryFilter
-  = FilterByAddress !(Set (Address Shelley))
-  | NoFilter
-  deriving (Eq, Show)
-
--- | Either a stake pool verification key hash or verification key file.
-data StakePoolVerificationKeyHashOrFile
-  = StakePoolVerificationKeyHash !(Hash StakePoolKey)
-  | StakePoolVerificationKeyFile !VerificationKeyFile
   deriving (Eq, Show)
