+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Governance.hs
@@ -15,6 +15,7 @@ import           Cardano.Api.TextView (TextViewDescription (..), textShow)
 import           Cardano.Api.Typed
 
 import           Cardano.CLI.Shelley.Parsers
+import           Cardano.CLI.Types
 
 import qualified Shelley.Spec.Ledger.TxData as Shelley
 
@@ -50,7 +51,6 @@ renderShelleyGovernanceError err =
 runGovernanceCmd :: GovernanceCmd -> ExceptT ShelleyGovernanceError IO ()
 runGovernanceCmd (GovernanceMIRCertificate mirpot vKeys rewards out) = runGovernanceMIRCertificate mirpot vKeys rewards out
 runGovernanceCmd (GovernanceUpdateProposal out eNo genVKeys ppUp) = runGovernanceUpdateProposal out eNo genVKeys ppUp
-runGovernanceCmd cmd = liftIO $ putStrLn $ "TODO: runGovernanceCmd: " ++ show cmd
 
 runGovernanceMIRCertificate
   :: Shelley.MIRPot
