+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Key.hs
@@ -35,6 +35,7 @@ import           Cardano.CLI.Byron.Key (CardanoEra (..))
 import qualified Cardano.CLI.Byron.Key as Byron
 import           Cardano.CLI.Helpers (textShow)
 import           Cardano.CLI.Shelley.Commands
+import           Cardano.CLI.Types (SigningKeyFile (..), VerificationKeyFile(..))
 
 
 data ShelleyKeyCmdError
