+++ cardano-cli/src/Cardano/CLI/Types.hs
@@ -2,9 +2,9 @@
 {-# LANGUAGE GeneralisedNewtypeDeriving #-}
 
 module Cardano.CLI.Types
-  ( CertificateFile (..)
-  , NodeAddress (..)
-  , NodeHostAddress (..)
+  ( CBORObject (..)
+  , CertificateFile (..)
+  , GenesisFile (..)
   , QueryFilter (..)
   , SigningKeyFile (..)
   , SocketPath (..)
@@ -15,29 +15,37 @@ module Cardano.CLI.Types
 
 import           Cardano.Prelude
 
-import           Data.IP (IP)
-import           Network.Socket (PortNumber)
+import           Data.Aeson
+import qualified Data.Text as Text
+
+import qualified Cardano.Chain.Slotting as Byron
 
 import           Cardano.Api.Typed
+
+-- | Specify what the CBOR file is
+-- i.e a block, a tx, etc
+data CBORObject = CBORBlockByron Byron.EpochSlots
+                | CBORDelegationCertificateByron
+                | CBORTxByron
+                | CBORUpdateProposalByron
+                | CBORVoteByron
+                deriving Show
 
 -- Encompasses stake certificates, stake pool certificates,
 -- genesis delegate certificates and MIR certificates.
 newtype CertificateFile = CertificateFile { unCertificateFile :: FilePath }
                           deriving newtype (Eq, Show)
 
--- | IPv4 address with a port number.
-data NodeAddress = NodeAddress
-  { naHostAddress :: !NodeHostAddress
-  , naPort :: !PortNumber
-  } deriving (Eq, Ord, Show)
+newtype GenesisFile = GenesisFile
+  { unGenesisFile :: FilePath }
+  deriving stock (Eq, Ord)
+  deriving newtype (IsString, Show)
 
--- Embedding a Maybe inside a newtype is somewhat icky but this seems to work
--- and removing the Maybe breaks the functionality in a subtle way that is difficult
--- to diagnose.
-newtype NodeHostAddress
-  = NodeHostAddress { unNodeHostAddress :: Maybe IP }
-  deriving newtype Show
-  deriving (Eq, Ord)
+instance FromJSON GenesisFile where
+  parseJSON (String genFp) = pure . GenesisFile $ Text.unpack genFp
+  parseJSON invalid = panic $ "Parsing of GenesisFile failed due to type mismatch. "
+                           <> "Encountered: " <> (Text.pack $ show invalid)
+
 
 -- | UTxO query filtering options.
 data QueryFilter
@@ -46,7 +54,7 @@ data QueryFilter
   deriving (Eq, Show)
 
 newtype SigningKeyFile = SigningKeyFile
-  { unSigningKeyFile ::  FilePath }
+  { unSigningKeyFile :: FilePath }
   deriving stock (Eq, Ord)
   deriving newtype (IsString, Show)
