+++ cardano-config/cardano-config.cabal
@@ -24,55 +24,13 @@ library
 
   exposed-modules:     Cardano.Config.Git.Rev
                        Cardano.Config.Git.RevFromGit
-                       Cardano.Config.Types
 
   build-depends:       base >=4.12 && <5
-                     , attoparsec
-                     , aeson
-                     , async
-                     , base16-bytestring
-                     , bytestring
-                     , canonical-json >= 0.6 && < 0.7
-                     , cardano-binary
-                     , cardano-crypto-class
-                     , cardano-crypto-wrapper
-                     , cardano-ledger
                      , cardano-prelude
-                     , cardano-slotting
-                     , cborg
-                     , containers
-                     , contra-tracer
-                     , cryptonite
-                     , directory
                      , file-embed
-                     , filepath
-                     , iohk-monitoring
-                     , lobemo-backend-aggregation
-                     , lobemo-backend-ekg
-                     , lobemo-backend-monitoring
-                     , lobemo-backend-trace-forwarder
-                     , network
-                     , network-mux
-                     , optparse-applicative
-                     , ouroboros-consensus
-                     , ouroboros-consensus-byron
-                     , ouroboros-consensus-cardano
-                     , ouroboros-consensus-mock
-                     , ouroboros-consensus-shelley
-                     , ouroboros-network
-                     , ouroboros-network-framework
                      , process
-                     , iproute
-                     , scientific
-                     , shelley-spec-ledger
-                     , string-conv
-                     , stm
                      , template-haskell
                      , text
-                     , transformers
-                     , transformers-except
-                     , typed-protocols
-                     , unordered-containers
 
   default-language:    Haskell2010
   default-extensions:  NoImplicitPrelude
