+++ cardano-node/src/Cardano/Node/Logging.hs
@@ -64,9 +64,7 @@ import           Cardano.BM.Trace (Trace, appendName, traceNamedObject)
 import qualified Cardano.BM.Trace as Trace
 
 import           Cardano.Config.Git.Rev (gitRev)
-import           Cardano.Config.Types (ConfigError (..), ViewMode (..))
-import           Cardano.Node.Types (ConfigYamlFilePath (..), NodeCLI (..), NodeConfiguration (..),
-                     parseNodeConfiguration)
+import           Cardano.Node.Types
 
 --------------------------------
 -- Layer
