+++ cardano-node/src/Cardano/Node/TUI/Run.hs
@@ -35,7 +35,6 @@ import           Cardano.Node.TUI.Drawing (LiveViewState (..), LiveViewThread (.
 import           Cardano.Node.TUI.EventHandler (LiveViewBackend (..))
 import           Cardano.Tracing.Peer (Peer (..))
 
-import           Cardano.Config.Types (NodeAddress (..))
 import           Cardano.Node.Types
 
 -- | Change a few fields in the LiveViewState after it has been initialized above.
