+++ cardano-node/test/Test/Cardano/Node/Json.hs
@@ -8,7 +8,7 @@ import           Cardano.Prelude
 
 import           Data.Aeson (decode, encode, fromJSON, toJSON)
 
-import           Cardano.Config.Types (NodeAddress (..), NodeHostAddress (..))
+import           Cardano.Node.Types (NodeAddress (..), NodeHostAddress (..))
 
 import           Hedgehog (Property, discover)
 import qualified Hedgehog
